refactor: extract open-source section into OpenSource pattern
Six pages (embed, screenshot, pdf, markdown, metadata, link-preview)
carried a copy-pasted ~300-line repo-cards section. It now lives in
components/patterns/OpenSource with repos/accent/caption as props,
mirroring the Timings pattern. getRepoStarsLabel is exported for the
JSON-LD interaction counters. Guard test keeps pages from re-declaring
the pattern locally.
