+++ src/components/patterns/OpenSource/index.js
@@ -1,0 +1,1 @@
+export { default, getRepoStarsLabel } from './OpenSource'
